pyproject.toml
[project]
name = "amc-peripheral"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.13",
    "beautifulsoup4>=4.13.3",
    "discord>=2.3.2",
    "django>=5.2.1",
    "django-ninja>=1.4.1",
    "elevenlabs>=1.56.0",
    "ffmpeg-normalize>=1.31.2",
    "gevent>=25.4.2",
    "google-cloud-texttospeech==2.27.0",
    "gunicorn>=23.0.0",
    "markdownify>=1.1.0",
    "mutagen>=1.47.0",
    "openai==1.96.1",
    "playwright>=1.50.0",
    "pycryptodome>=3.22.0",
    "pydantic>=2.10.6",
    "pytz>=2025.2",
    "requests>=2.32.3",
    "ruff>=0.11.2",
    "savconverter>=0.1.1",
    "the-retry>=0.1.1",
    "urllib3>=2.2.3",
    "uvicorn-worker>=0.3.0",
    "uvicorn[standard]>=0.34.2",
    "websockets>=15.0",
    "yt-dlp==2025.11.1.232827.dev0",
    "yt-dlp-ejs>=0.3.0",
]

[project.scripts]
motortown_bot = "amc_peripheral.bot.bot:main"
motortown_radio = "amc_peripheral.radio.radio:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
  "amc_peripheral"
]

[dependency-groups]
dev = [
    "ruff>=0.11.2",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "dpytest>=0.7.0",
    "aioresponses>=0.7.6",
]
